Recursively create, go to, leave, and delete directories in C in UNIX environment
Homework #6 (problem 4.16, pg. 142 in Advanced Programming in the UNIX
Environment by Richard Stevens), asked the following: "Does the UNIX
System have a fundamental limitation on the depth of a directory tree?
To find out, write a program that creates a directory and then changes
to that directory, in a loop.  Make certain that the length of the
absolute pathname of the leaf of this directory is greater than your
system's PATH_MAX limit."

This assignment, Homework #7, was an extension of Homework #6.  It
required that we add the functionality to delete the created directories
in all cases, regardless of the manner in which the program terminates.